backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
pydub==0.25.1
numpy==1.24.3
torch==2.1.0
transformers==4.35.0
speechrecognition==3.10.0
pyttsx3==2.90
whisper-openai==20231117
ollama==0.1.7
asterisk-agi==0.3.1
redis==5.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
